KA-160 Rename model
